feat: upload alterations related to looking for Java collection-like libs' import
Adds logic to look for some Java collection-like libs that would probably account for false positives.
This logic only works if passed the flag -checkcollection to the operator-search script.
Also, fix the count for files by adding the division for the total of operators inspected. This fix is
necessary since the same file is replicated to operators workers(gorotines) that counts separately for
the presence of each operator.